requirements.txt
requests~=2.32.3
humanize~=4.10.0
protobuf~=5.27.3
google-auth-oauthlib~=1.2.1
google-api-python-client~=2.142.0
tqdm~=4.66.5